Move denied users functions from vettingModel to their own deniedUsersModel
