+++ app/database/models/vettingModel.js
@@ -42,23 +42,6 @@ const approveUser = async (sub) => {
   return newUser;
 };
 
-// copies user to denied_users table and deletes them from vetting table
-const denyUser = async (sub) => {
-  const user = await findVettingUserBySub(sub);
-  const deniedUserId = await db('denied_users').insert(user);
-  deleteUser(sub);
-  return deniedUserId;
-};
-
-const findDeniedUsers = async () => {
-  db('denied_users');
-};
-
-const findDeniedUserBySub = async (sub) => {
-  const user = await db('denied_users').where({ sub }).first();
-  return user;
-};
-
 module.exports = {
   addVettingUser,
   findVettingUserBySub,
@@ -66,7 +49,4 @@ module.exports = {
   approveUser,
   deleteUser,
   findAll,
-  denyUser,
-  findDeniedUserBySub,
-  findDeniedUsers,
 };
